CMakeClean: Support the -B flag for --clear and --clean

+++ Modules/CMakeClean/CMakeClean.cpp
@@ -1,6 +1,7 @@
 #include "HookDll.hpp"
 
 #include <cstdio>
+#include <string>
 
 HOOK(Shell32.dll, HINSTANCE WINAPI, ShellExecuteA)(
 	HWND   hwnd,
@@ -15,16 +16,64 @@ HOOK(Shell32.dll, HINSTANCE WINAPI, ShellExecuteA)(
 	return original_ShellExecuteA(hwnd, lpOperation, lpFile, lpParameters, lpDirectory, nShowCmd);
 }
 
+static bool FileExists(const wchar_t* szFileName)
+{
+	return GetFileAttributesW(szFileName) != INVALID_FILE_ATTRIBUTES;
+};
+
+static wchar_t szCurrentDirectory[MAX_PATH * 10];
+
 HOOK_ENTRYPOINT()
 {
 	dlog();
+
+	bool restoreCurrentDirectory = false;
+	GetCurrentDirectoryW(_countof(szCurrentDirectory), szCurrentDirectory);
+
 	auto commandLine = GetCommandLineW();
+	int argc = 0;
+	auto argv = CommandLineToArgvW(commandLine, &argc);
+	if (argv)
+	{
+		for (int i = 1; i < argc; i++)
+		{
+			auto arg = argv[i];
+			if (wcsstr(arg, L"-B") == arg)
+			{
+				const wchar_t* buildDir = nullptr;
+				if (wcslen(arg) == 2)
+				{
+					buildDir = argv[i + 1];
+				}
+				else if (i + 1 < argc)
+				{
+					buildDir = arg + 2;
+				}
+
+				if (buildDir != nullptr)
+				{
+					dlogp("SetCurrentDirectory: %S", buildDir);
+					SetCurrentDirectoryW(buildDir);
+					restoreCurrentDirectory = true;
+				}
+			}
+		}
+		LocalFree(argv);
+	}
+	{
+		auto buildDir = wcsstr(commandLine, L"-B");
+		if (buildDir)
+		{
+			// Skip -B
+			buildDir += 2;
+			// Skip spaces
+			while (*buildDir == L' ')
+				buildDir++;
+		}
+	}
+
 	if (wcsstr(commandLine, L" --clean"))
 	{
-		auto FileExists = [](const wchar_t* szFileName)
-		{
-			return GetFileAttributesW(szFileName) != INVALID_FILE_ATTRIBUTES;
-		};
 		bool cacheDeleted = true;
 		if (FileExists(L"CMakeCache.txt"))
 		{
@@ -46,11 +95,35 @@ HOOK_ENTRYPOINT()
 	}
 	else if (wcsstr(commandLine, L"--clear"))
 	{
-		// TODO: nicer error handling
-		// Thanks to Jonas for the help with the command
-		system("rmdir /s /q . > nul 2>&1 & dir /b");
-		return 0;
+		if (FileExists(L"CMakeCache.txt"))
+		{
+			// TODO: nicer error handling
+			// Thanks to Jonas for the help with the command
+			system("rmdir /s /q . > nul 2>&1 & dir /b");
+		}
+
+		if (restoreCurrentDirectory)
+		{
+			// Remove the --clear flag and continue execution
+			std::wstring cleanCommandLine(commandLine);
+			auto clearIdx = cleanCommandLine.find(L"--clear");
+			auto hasSpace = cleanCommandLine.size() > clearIdx + 7 && cleanCommandLine[clearIdx + 7] == L' ';
+			cleanCommandLine = cleanCommandLine.erase(clearIdx, 7 + hasSpace ? 1 : 0);
+			dlogp("commandLine: %S", cleanCommandLine.c_str());
+			wcscpy(commandLine, cleanCommandLine.c_str());
+		}
+		else
+		{
+			// Fini
+			return 0;
+		}
 	}
+
+	if (restoreCurrentDirectory)
+	{
+		SetCurrentDirectoryW(szCurrentDirectory);
+	}
+
 	return original_EntryPoint();
 }
